refactor(pagination): rename .scss files to .module.scss
