+++ packages/pagination/src/Pagination.module.scss
