+++ packages/pagination/src/TruncateIndicator.tsx
@@ -1,7 +1,7 @@
 import React from "react"
 import ellipsis from "@kaizen/component-library/icons/ellipsis.icon.svg"
 import { Icon } from "@kaizen/component-library"
-import styles from "./Pagination.scss"
+import styles from "./Pagination.module.scss"
 
 export const TruncateIndicator = () => (
   <div
